change a elements to buttons
this change replaces the navlist and accordion list `a` elements with
`button` elements instead. this is being done to solve some issues that arise
when embedding the camgi output within an iframe or other type of
wrapper. also, when reviewing the html standard[0] it appears that
`a` elements may not be contextually appropriate for what camgi does
since these are not hyperlinks to other pages.

[0]
https://html.spec.whatwg.org/multipage/text-level-semantics.html#the-a-element