mustgather: Use clearer+more efficient loop pattern
When to use combinators in Rust involved a lot of learning over
time for me.  Sometimes, combinators are a lot clearer; other
times they aren't.

In this case, I think it's much clearer to directly perform
operations inside the loop.  This is also more efficient because
we aren't calling `.collect()` to create an intermediate vector
(though we could have just dropped that call at the end of the
 combinator chain).

Also of note instead of `is_ok()` + `unwrap()` we can just call `ok()`.

That said, I personally think we should be making use of
`Result<>` here to propagate errors instead of ignoring them.