add clusteroperators to output
this change brings in the clusteroperator type to the output, it will
add an error status to an record which has the available condition set
to false or the degraded condition set to true. further, it will add the
warning status to any record which contains the upgradeable condition
set to false.

this change also fixes an issue with warnings overriding errors in the
generated accordion output.